+++ src/server/mod.rs
@@ -1,4 +1,5 @@
 //! Server module
 
+#[cfg_attr(docsrs, doc(cfg(feature = "axum")))]
 #[cfg(feature = "axum")]
 pub mod axum;